Theme is has changed! There are multiple changes

+++ app/src/main/java/com/example/liderstoreagent/data/models/discountsmodel/Discounts.kt
@@ -2,7 +2,7 @@ package com.example.liderstoreagent.data.models.discountsmodel
 
 data class Discounts(
     val deadline: String?,
-    val discount: Int?,
+    val discount: Double?,
     val id: Int,
     val name: String
 )
